Refactor module state management: update to use CurrentModuleSlice and add moduleId retrieval in KanbanView and ListView
