+++ client/src/components/Team/ManageWork/Views/KanbanView.js
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react'
-import { Link } from 'react-router-dom';
+import { Link, useParams } from 'react-router-dom';
 import utils from '../../../../utils';
 import './KanbanView.css'
-import { setCurrentModuleView } from '../../../../store/reducers/CurrentModuleViewSlice';
+import { setCurrentModuleId, setCurrentModuleView } from '../../../../store/reducers/CurrentModuleSlice';
 import { useDispatch } from 'react-redux';
 
 const RowComponent = ({ item }) => {
@@ -51,9 +51,11 @@ const ColumnComponent = ({ column }) => {
 const KanbanView = ({ ManageWorkListKanban }) => {
 
     const dispatch = useDispatch();
+    const { moduleId } = useParams();
 
     useEffect(() => {
         dispatch(setCurrentModuleView("kanban"));
+        dispatch(setCurrentModuleId(moduleId));
     }, []);
 
     return (
